Adds env check & messages to destroy script.
Adds a message to indicate that there is still an AppStream image that
needs to be removed manually.

+++ scripts/destroy.js
@@ -10,6 +10,31 @@ const exec = (command, options={}) => {
   execSync(command, combinedOptions);
 };
 
+const { HORTA_ORG, HORTA_STAGE, AWS_REGION, AWS_ACCOUNT } =
+  process.env;
+console.log(chalk.cyan("🔎 Checking environment."));
+
+const expectedEnvVars = [
+  "HORTA_ORG",
+  "HORTA_STAGE",
+  "AWS_REGION",
+  "AWS_ACCOUNT"
+];
+
+let missingVarsCount = 0;
+
+expectedEnvVars.forEach(envVar => {
+  if (!process.env[envVar]) {
+    console.log(chalk.red(`🚨 Environment variable ${envVar} was not set.`));
+    missingVarsCount += 1;
+  }
+});
+
+if (missingVarsCount > 0) {
+  process.exit(1);
+}
+console.log(chalk.green("✅ environment looks good."));
+
 console.log(chalk.red("🚨 Removing web admin frontend stack."));
 exec(
   `npm run cdk -- destroy -f --require-approval never -c deploy=admin_website`,
@@ -33,3 +58,8 @@ exec(
   `npm run cdk -- destroy -f --all --require-approval never`,
   { cwd: "./vpc_stack/" }
 );
+
+console.log(chalk.green("✅ All stacks have been removed."));
+const imageName = `${HORTA_ORG}-hc-HortaCloudWorkstation-${HORTA_STAGE}`;
+console.log(chalk.yellow(`⚠️  Please manually remove the AppStream image ${imageName} at:`));
+console.log(chalk.white("https://console.aws.amazon.com/appstream2/home"));
